+++ packages/mask/src/plugins/NFTCard/SNSAdaptor/components/NFTCardContent.tsx
@@ -6,9 +6,9 @@ import { WalletMessages } from '../../../Wallet/messages.js'
 import { NetworkPluginID } from '@masknet/web3-shared-base'
 import { NFTCardDialogTabs } from './NFTCardDialog.js'
 import { useStyles } from '../../useStyles.js'
-import { AboutTab } from '../Tabs/AboutTab.js'
-import { OffersTab } from '../Tabs/OffersTab.js'
-import { ActivitiesTab } from '../Tabs/ActivitiesTab.js'
+import { AboutTab } from './tabs/AboutTab.js'
+import { OffersTab } from './tabs/OffersTab.js'
+import { ActivitiesTab } from './tabs/ActivitiesTab.js'
 import { NFTBasicInfo } from '../../../../components/shared/NFTCard/NFTBasicInfo.js'
 import { LoadingBase } from '@masknet/theme'
 import { base as pluginDefinition } from '../../base.js'
