+++ packages/mask/src/plugins/NFTCard/SNSAdaptor/components/NFTCardDialog.tsx
@@ -12,9 +12,9 @@ import { CrossIsolationMessages } from '@masknet/shared-base'
 import { TabContext } from '@mui/lab'
 import { DialogContent, Tab } from '@mui/material'
 import { NFTCardContent } from './NFTCardContent.js'
-import { useStyles } from '../useStyles.js'
-import { useI18N } from '../../../utils/index.js'
-import { Context } from './hooks/useContext.js'
+import { useStyles } from '../../useStyles.js'
+import { useI18N } from '../../../../utils/index.js'
+import { Context } from '../hooks/useContext.js'
 
 export enum NFTCardDialogTabs {
     About = 'About',
