+++ packages/mask/src/plugins/NFTCard/SNSAdaptor/components/NFTInfoCard.tsx
@@ -4,7 +4,7 @@ import { Link, Typography } from '@mui/material'
 import type { Web3Helper } from '@masknet/plugin-infra/web3'
 import { SourceType } from '@masknet/web3-shared-base'
 import { useWeb3State } from '@masknet/plugin-infra/web3'
-import { useI18N } from '../../../utils/index.js'
+import { useI18N } from '../../../../utils/index.js'
 
 const useStyles = makeStyles()((theme) => ({
     wrapper: {
