+++ packages/mask/src/plugins/NFTCard/SNSAdaptor/components/NFTPropertiesCard.tsx
@@ -2,7 +2,7 @@ import { makeStyles } from '@masknet/theme'
 import { Typography } from '@mui/material'
 import type { Web3Helper } from '@masknet/plugin-infra/web3'
 import { NFTRank } from './NFTRank.js'
-import { useI18N } from '../../../utils/index.js'
+import { useI18N } from '../../../../utils/index.js'
 
 const useStyles = makeStyles()((theme) => ({
     wrapper: {
