+++ packages/mask/src/plugins/NFTCard/types.ts
@@ -1,0 +1,8 @@
+export enum ActivityType {
+    Transfer = 'Transfer',
+    Mint = 'Mint',
+    Sale = 'Sale',
+    Offer = 'Offer',
+    List = 'List',
+    CancelOffer = 'Cancel Offer',
+}
